requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
db-dtypes==1.2.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
numpy==1.26.4
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
